# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-partyassets
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "4.0"
servers:
- url: http://localhost:52333
  description: Generated server url
tags:
- name: Metadata - Statusreasons
  description: Statusreasons operations
paths:
  /{municipalityId}/metadata/statusreasons/{status}:
    get:
      tags:
      - Metadata - Statusreasons
      summary: Get status reasons
      operationId: readReasons
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: status
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Status"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Metadata - Statusreasons
      summary: Create status reasons
      operationId: createReasons
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: status
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Status"
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        "201":
          description: Created - Successful operation
          headers:
            Location:
              description: Location of the created resource.
              style: simple
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Metadata - Statusreasons
      summary: Delete status reasons
      operationId: deleteReasons
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: status
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Status"
      responses:
        "204":
          description: No content - Successful operation
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/assets:
    get:
      tags:
      - Assets
      summary: Get assets
      operationId: getAssets
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: assetId
        in: query
        description: External asset id (e.g. PRH-123456789) used as an identifier
          by external systems
        required: false
        schema:
          type: string
          description: External asset id (e.g. PRH-123456789) used as an identifier
            by external systems
          examples:
          - PRH-123456789
      - name: origin
        in: query
        description: Source of origin for the asset
        required: false
        schema:
          type: string
          description: Source of origin for the asset
          examples:
          - CASEDATA
      - name: partyId
        in: query
        description: PartyId
        required: true
        schema:
          type: string
          description: PartyId
          examples:
          - 123e4567-e89b-12d3-a456-426614174000
      - name: type
        in: query
        description: Asset type
        required: false
        schema:
          type: string
          description: Asset type
          examples:
          - PERMIT
      - name: issued
        in: query
        description: Issued date
        required: false
        schema:
          type: string
          format: date
          description: Issued date
          examples:
          - 2021-01-01
      - name: validTo
        in: query
        description: Valid to date
        required: false
        schema:
          type: string
          format: date
          description: Valid to date
          examples:
          - 2021-12-31
      - name: status
        in: query
        description: Asset status
        required: false
        schema:
          $ref: "#/components/schemas/Status"
      - name: statusReason
        in: query
        description: Status reason
        required: false
        schema:
          type: string
          description: Status reason
          examples:
          - Status reason
      - name: description
        in: query
        description: Asset description
        required: false
        schema:
          type: string
          description: Asset description
          examples:
          - Asset description
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Assets
      summary: Create an asset
      operationId: createAsset
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetCreateRequest"
        required: true
      responses:
        "201":
          description: Created - Successful operation
          headers:
            Location:
              description: Location of the created resource.
              style: simple
              schema:
                type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/assets/{id}:
    get:
      tags:
      - Assets
      summary: Get asset
      operationId: getAsset
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - Assets
      summary: Delete an asset
      operationId: deleteAsset
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No content - Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      tags:
      - Assets
      summary: Update an asset
      operationId: updateAsset
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdateRequest"
        required: true
      responses:
        "204":
          description: No content - Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/metadata/statusreasons:
    get:
      tags:
      - Metadata - Statusreasons
      summary: Get all status reasons
      operationId: readAllReasons
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api-docs:
    get:
      tags:
      - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Status:
      type: string
      description: Status model
      enum:
      - ACTIVE
      - EXPIRED
      - BLOCKED
      - TEMPORARY
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause: {}
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    AssetCreateRequest:
      type: object
      properties:
        assetId:
          type: string
          description: External asset id (e.g. PRH-123456789) used as an identifier
            by external systems
          examples:
          - PRH-123456789
        origin:
          type: string
          description: Source of origin for the asset
          examples:
          - CASEDATA
        partyId:
          type: string
          description: PartyId
          examples:
          - 123e4567-e89b-12d3-a456-426614174000
        caseReferenceIds:
          type: array
          description: Case reference ids
          examples:
          - "[\"123e4567-e89b-12d3-a456-426614174000\"]"
          items:
            type: string
        type:
          type: string
          description: Asset type
          examples:
          - PERMIT
          minLength: 1
        issued:
          type: string
          format: date
          description: Issued date
          examples:
          - 2021-01-01
        validTo:
          type: string
          format: date
          description: Valid to date
          examples:
          - 2021-12-31
        status:
          $ref: "#/components/schemas/Status"
          description: Asset status
          examples:
          - ACTIVE
        statusReason:
          type: string
          description: Status reason
          examples:
          - Status reason
        description:
          type: string
          description: Asset description
          examples:
          - Asset description
        additionalParameters:
          type: object
          additionalProperties:
            type: string
          description: Additional parameters
          examples:
          - "{\"foo\":\"bar\"}"
        jsonParameters:
          type: array
          description: JSON parameters
          items:
            $ref: "#/components/schemas/AssetJsonParameter"
      required:
      - issued
      - partyId
      - status
      - type
    AssetJsonParameter:
      type: object
      properties:
        key:
          type: string
          description: Parameter key
          examples:
          - personParameter
          minLength: 1
        value:
          $ref: "#/components/schemas/JsonNode"
          description: Parameter value with the JSON structure
          example:
            firstName: Joe
            lastName: Doe
        schemaId:
          type: string
          description: Schema ID
          examples:
          - person_1.0
          minLength: 1
      required:
      - key
      - schemaId
      - value
    JsonNode: {}
    AssetUpdateRequest:
      type: object
      properties:
        caseReferenceIds:
          type: array
          description: Case reference ids
          examples:
          - "[\"123e4567-e89b-12d3-a456-426614174000\"]"
          items:
            type: string
        validTo:
          type: string
          format: date
          description: Valid to date
          examples:
          - 2021-12-31
        status:
          $ref: "#/components/schemas/Status"
          description: Asset status
          examples:
          - ACTIVE
        statusReason:
          type: string
          description: Status reason
          examples:
          - Status reason
        additionalParameters:
          type: object
          additionalProperties:
            type: string
          description: Additional parameters
          examples:
          - "{\"foo\":\"bar\"}"
        jsonParameters:
          type: array
          description: JSON parameters
          items:
            $ref: "#/components/schemas/AssetJsonParameter"
    Asset:
      type: object
      properties:
        id:
          type: string
          description: Unique id of asset
          examples:
          - 1c8f38a6-b492-4037-b7dc-de5bc6c629f0
        assetId:
          type: string
          description: External asset id (e.g. PRH-123456789) used as an identifier
            by external systems
          examples:
          - PRH-123456789
        origin:
          type: string
          description: Source of origin for the asset
          examples:
          - CASEDATA
        partyId:
          type: string
          description: PartyId
          examples:
          - 123e4567-e89b-12d3-a456-426614174000
        caseReferenceIds:
          type: array
          description: Case reference ids
          examples:
          - "[\"945576d3-6e92-4118-ba33-53582d338ad3\"]"
          items:
            type: string
        type:
          type: string
          description: Asset type
          examples:
          - PERMIT
        issued:
          type: string
          format: date
          description: Issued date
          examples:
          - 2021-01-01
        validTo:
          type: string
          format: date
          description: Valid to date
          examples:
          - 2021-12-31
        status:
          $ref: "#/components/schemas/Status"
          description: Asset status
          examples:
          - ACTIVE
        statusReason:
          type: string
          description: Status reason
          examples:
          - Status reason
        description:
          type: string
          description: Asset description
          examples:
          - Asset description
        additionalParameters:
          type: object
          additionalProperties:
            type: string
          description: Additional parameters
          examples:
          - "{\"foo\":\"bar\"}"
        jsonParameters:
          type: array
          description: JSON parameters
          items:
            $ref: "#/components/schemas/AssetJsonParameter"
  securitySchemes: {}